[data-literate old][m]: Rename new data literate folder
