+++ examples/data-literate/lib/markdown.js
