+++ examples/data-literate/package.json
@@ -8,34 +8,29 @@
     "export": "next export",
     "start": "next start"
   },
-  "engines": {
-    "node": ">12"
-  },
   "dependencies": {
     "@headlessui/react": "^1.3.0",
     "@heroicons/react": "^1.0.3",
-    "@mdx-js/loader": "^2.0.0",
-    "@mdx-js/react": "^2.0.0",
-    "@next/mdx": "^12.1.0",
-    "@tailwindcss/typography": "^0.5.2",
-    "autoprefixer": "^10.4.2",
+    "@mdx-js/loader": "^1.6.22",
+    "@tailwindcss/typography": "^0.4.0",
+    "autoprefixer": "^10.0.4",
     "frictionless.js": "^0.13.4",
+    "gray-matter": "^4.0.3",
     "next": "12.1.0",
+    "next-mdx-remote": "^3.0.4",
     "papaparse": "^5.3.1",
-    "portal": "https://github.com/datopian/portal.js.git",
-    "postcss": "^8.4.7",
+    "postcss": "^8.2.10",
     "prop-types": "^15.7.2",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "react-vega": "^7.4.4",
     "remark": "^13.0.0",
     "remark-footnotes": "^3.0.0",
-    "remark-frontmatter": "^4.0.1",
     "remark-gfm": "^1.0.0",
-    "remark-mdx-frontmatter": "^1.1.1",
+    "remark-html": "^13.0.2",
     "remark-slug": "^6.1.0",
     "remark-toc": "^7.2.0",
-    "tailwindcss": "^3.0.23",
+    "tailwindcss": "^2.2.16",
     "vega": "^5.20.2",
     "vega-lite": "^5.1.0",
     "xlsx": "^0.17.0"
