+++ examples/data-literate/pages/[...slug].js
