+++ examples/data-literate/pages/_app.js
@@ -1,12 +1,8 @@
 import '../styles/globals.css'
 import '../styles/tailwind.css'
-import DataLiterate from '../components/DataLiterate'
-
 
 function MyApp({ Component, pageProps }) {
-  return (
-    <DataLiterate children={{ Component, pageProps }}/>
-  )
+  return <Component {...pageProps} />
 }
 
 export default MyApp
