+++ examples/data-literate/tailwind.config.js
@@ -1,8 +1,9 @@
 const defaultTheme = require("tailwindcss/defaultTheme");
 
 module.exports = {
+  mode: 'jit',
   // purge: ['./pages/**/*.{js,ts,jsx,tsx}', './components/**/*.{js,ts,jsx,tsx}'],
-  content: [
+  purge: [
     "./pages/**/*.js",
     "./pages/**/*.ts",
     "./pages/**/*.jsx",
@@ -22,6 +23,9 @@ module.exports = {
         mono: ["Inconsolata", ...defaultTheme.fontFamily.mono]
       }
     },
+  },
+  variants: {
+    extend: {},
   },
   plugins: [
     require('@tailwindcss/typography'),
